MAJOR UPDATE CUTLINE!
I wanted the code to be easier to clean up. So now that I started making the simplifying assumption that the code must be checked out to ~/Blue2thprinting, I can know where paths like Scripts or bluez-5.66 will be (instead of just using where they were on my first set up machine.)

So this changes a bunch of paths to the new assumptions. It also updates the default assumed username from "pi" to "user" (my typical class default), because as of right now the full code doesn't work on Raspberry Pis (due to the Braktooth dependency on x86 binaries.) (not that this matters really since the script fixes it up either way now.)

I'm still working through all the changes, but this commit marks the cutover for most paths (and at least the basic functionality w/o Swyen/Brak is confirmed working in a VM.)